Added Role Selector close functionality when clicking outside of it (#211)
* Added dropdown close functionality when clicking
outside.

Added an event listener to the document that checks if the user clicks
outside the dropdown and closes it if so. This behavior is implemented
using a useRef hook and a mousedown event. The cleanup function is also
added to remove the event listener when the component unmounts.

* Optimized event listener in RoleSelector component
for improved performance.
Updated event listener for RoleSelector component
to only be added when the dropdown is opened and removed
when the dropdown is closed.

* Removed package-lock.json and yarn.lock

* Revert "Removed package-lock.json and yarn.lock"

This reverts commit f9dbf056b6003377d5bcf4f3a731afe7324f000b.

* Remove yarn.lock, package-lock.json from PR

* revert yarn.lock

* optimisation

* change React.useRef to useRef

---------

Co-authored-by: Jing Hua <tohjinghua123@gmail.com>